# feed-service/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Feed API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/hello/hello_world": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get if server is up",
                "tags": [
                    "Hello World"
                ],
                "summary": "Greets with Hello World",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/pinboards/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new pin board",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Pin Boards"
                ],
                "summary": "Create new pin board for the user",
                "parameters": [
                    {
                        "description": "Pin Board Data",
                        "name": "pinboard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PinBoard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.PinBoard": {
            "type": "object",
            "required": [
                "board_id",
                "board_name"
            ],
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "board_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "is_public": {
                    "type": "boolean"
                },
                "is_updated": {
                    "type": "boolean"
                },
                "likes_count": {
                    "type": "integer"
                },
                "share_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}